SEQNG-478 Code cleanups
