+++ modules/edu.gemini.seqexec.web/edu.gemini.seqexec.web.client/src/main/scala/edu/gemini/seqexec/web/client/components/sequence/steps/StepsTable.scala
@@ -5,7 +5,7 @@ package edu.gemini.seqexec.web.client.components.sequence.steps
 
 import scala.scalajs.js
 import diode.react.ModelProxy
-import edu.gemini.seqexec.model.Model.{Instrument, Step, StepType, StepState}
+import edu.gemini.seqexec.model.Model.{Instrument, Step}
 // import edu.gemini.seqexec.web.client.ModelOps._
 // import edu.gemini.seqexec.web.client.model.Pages.{SeqexecPages, SequenceConfigPage}
 import edu.gemini.seqexec.web.client.model.Pages.SeqexecPages
@@ -14,11 +14,10 @@ import edu.gemini.seqexec.web.client.model.Pages.SeqexecPages
 import edu.gemini.seqexec.web.client.circuit.{ClientStatus, StepsTableFocus}
 import edu.gemini.seqexec.web.client.components.SeqexecStyles
 import edu.gemini.seqexec.web.client.components.sequence.steps.OffsetFns._
-import edu.gemini.seqexec.web.client.lenses.stepTypeO
-import edu.gemini.seqexec.web.client.semanticui.{Size => SSize}
+// import edu.gemini.seqexec.web.client.lenses.stepTypeO
 // import edu.gemini.seqexec.web.client.semanticui.elements.icon.Icon
 // import edu.gemini.seqexec.web.client.semanticui.elements.icon.Icon._
-import edu.gemini.seqexec.web.client.semanticui.elements.label.Label
+// import edu.gemini.seqexec.web.client.semanticui.elements.label.Label
 // import edu.gemini.seqexec.web.client.semanticui.elements.message.IconMessage
 // import edu.gemini.seqexec.web.client.services.HtmlConstants.iconEmpty
 import japgolly.scalajs.react._
@@ -30,8 +29,8 @@ import japgolly.scalajs.react.vdom.html_<^._
 import scalacss.ScalaCssReact._
 import scalaz.std.AllInstances._
 import scalaz.syntax.foldable._
-import scalaz.syntax.equal._
-import scalaz.syntax.show._
+// import scalaz.syntax.equal._
+// import scalaz.syntax.show._
 // import scalaz.syntax.std.boolean._
 import scalaz.syntax.std.option._
 import react.virtualized._
@@ -47,67 +46,6 @@ object ColWidths {
   val ObjectTypeWidth: Int = 100
 }
 
-/**
-  * Component to display the step id
-  */
-object StepIdCell {
-  private val component = ScalaComponent.builder[Int]("StepIdCell")
-    .stateless
-    .render_P { p =>
-      <.div(
-        s"${p + 1}")
-    }.build
-
-  def apply(i: Int): Unmounted[Int, Unit, Unit] = component(i)
-}
-
-/**
-  * Component to display the offsets
-  */
-object OffsetsDisplayCell {
-  final case class Props(offsetsDisplay: OffsetsDisplay, step: Step)
-
-  private val component = ScalaComponent.builder[Props]("OffsetsDisplayCell")
-    .stateless
-    .render_P { p =>
-      <.div( // Column step offset
-        p.offsetsDisplay match {
-          case OffsetsDisplay.DisplayOffsets(offsetWidth) =>
-            OffsetBlock(OffsetBlock.Props(p.step, offsetWidth))
-          case _ => EmptyVdom
-        }
-      )
-    }.build
-
-  def apply(i: Props): Unmounted[Props, Unit, Unit] = component(i)
-}
-
-/**
-  * Component to display the object type
-  */
-object ObjectTypeCell {
-  private val component = ScalaComponent.builder[Step]("ObjectTypeCell")
-    .stateless
-    .render_P { p =>
-      <.div( // Column object type
-        SeqexecStyles.rightCell,
-        stepTypeO.getOption(p).map { st =>
-          val stepTypeColor = st match {
-            case _ if p.status === StepState.Completed => "light gray"
-            case StepType.Object                          => "green"
-            case StepType.Arc                             => "violet"
-            case StepType.Flat                            => "grey"
-            case StepType.Bias                            => "teal"
-            case StepType.Dark                            => "black"
-            case StepType.Calibration                     => "blue"
-          }
-          Label(Label.Props(st.shows, color = stepTypeColor.some, size = SSize.Small))
-        }.whenDefined
-      )
-    }.build
-
-  def apply(i: Step): Unmounted[Step, Unit, Unit] = component(i)
-}
 
 /**
   * Container for a table with the steps
